third_party/blink/web_tests/crypto/subtle/deserialize-legacy-aes-key-empty-usages-expected.txt

Check that an already imported key with empty usages can be deserialized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS key.type is "secret"
PASS key.extractable is true
PASS key.algorithm.name is "AES-CBC"
PASS key.usages is []
PASS successfullyParsed is true

TEST COMPLETE
